Perspective Puzzle Base
The base things about the perspective puzzle is ready. Just the clock or the time console needs a sript which lets it change the x value of the DirectionalLight.

Basically: In the room, there are thousands of different shapes and objects. The player won't understand what's going on. And there's a console in front of the exit door which needs a 4 digit code to open.
In the middle of the place, there's a clock. The player can change the time from it.
They need to change it to the 2 required values to find the first and second half of the four digit code. Afterwards, they will combine them; write it down to the console and open the door to exit from the episode.

I'm not sure about which eoisode this puzzle should be. We can discuss it later.